Cleaned up formatting - made consistent across all classes. No functional changes.

+++ src/java/jnibwapi/model/ChokePoint.java
@@ -1,4 +1,5 @@
 package jnibwapi.model;
+
 /**
  * Represents a choke point in a StarCraft map.
  * 
@@ -20,11 +21,11 @@ public class ChokePoint {
 	private int secondSideY;
 	private Region firstRegion;
 	private Region secondRegion;
-
+	
 	public ChokePoint(int[] data, int index) {
 		centerX = data[index++];
 		centerY = data[index++];
-		radius = ((double)data[index++])/fixedScale;
+		radius = data[index++] / fixedScale;
 		firstRegionID = data[index++];
 		secondRegionID = data[index++];
 		firstSideX = data[index++];
@@ -36,55 +37,55 @@ public class ChokePoint {
 	public Region getOtherRegion(Region region) {
 		return region.equals(firstRegion) ? secondRegion : firstRegion;
 	}
-
+	
 	public Region getFirstRegion() {
 		return firstRegion;
 	}
-
+	
 	public void setFirstRegion(Region firstRegion) {
 		this.firstRegion = firstRegion;
 	}
-
+	
 	public Region getSecondRegion() {
 		return secondRegion;
 	}
-
+	
 	public void setSecondRegion(Region secondRegion) {
 		this.secondRegion = secondRegion;
 	}
-
+	
 	public int getCenterX() {
 		return centerX;
 	}
-
+	
 	public int getCenterY() {
 		return centerY;
 	}
-
+	
 	public double getRadius() {
 		return radius;
 	}
-
+	
 	public int getFirstRegionID() {
 		return firstRegionID;
 	}
-
+	
 	public int getSecondRegionID() {
 		return secondRegionID;
 	}
-
+	
 	public int getFirstSideX() {
 		return firstSideX;
 	}
-
+	
 	public int getFirstSideY() {
 		return firstSideY;
 	}
-
+	
 	public int getSecondSideX() {
 		return secondSideX;
 	}
-
+	
 	public int getSecondSideY() {
 		return secondSideY;
 	}
